% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2slurm-package.R
\docType{package}
\name{r2slurm-package}
\alias{r2slurm}
\alias{r2slurm-package}
\title{r2slurm: Create and Submit Slurm Jobs from R}
\description{
Provides a chainable interface to create, render, and submit Slurm jobs from R.
}
\author{
\strong{Maintainer}: Joshua Kaluf \email{jkaluf2@illinois.edu}

}
\keyword{internal}
