% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_sbatch}
\alias{render_sbatch}
\title{Render SBATCH header lines}
\usage{
render_sbatch(job)
}
\arguments{
\item{job}{\code{slurm_job} object.}
}
\value{
Character vector of \code{#SBATCH} lines.
}
\description{
Converts a named list of Slurm options into \code{#SBATCH} directives
suitable for inclusion at the top of a submission script.
}
\examples{
render_sbatch(list(job_name = "test", cpus = 4, mem = "8G"))

}
