% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_cpus}
\alias{set_cpus}
\title{Set CPUs per task}
\usage{
set_cpus(job, cpus)
}
\arguments{
\item{job}{A \code{slurm_job} object.}

\item{cpus}{Number of CPUs.}
}
\value{
Updated \code{slurm_job} object.
}
\description{
Set CPUs per task
}
\examples{
job <- slurm_job("echo Hello") \%>\% set_cpus(4)

}
