% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_job_name}
\alias{set_job_name}
\title{Set Slurm job name}
\usage{
set_job_name(job, name)
}
\arguments{
\item{job}{A \code{slurm_job} object.}

\item{name}{Job name.}
}
\value{
Updated \code{slurm_job} object.
}
\description{
Set Slurm job name
}
\examples{
job <- slurm_job("echo Hello") \%>\% set_job_name("my_job")

}
