% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_mem}
\alias{set_mem}
\title{Set memory for a Slurm job}
\usage{
set_mem(job, mem)
}
\arguments{
\item{job}{A \code{slurm_job} object.}

\item{mem}{Memory value (e.g. "16G").}
}
\value{
Updated \code{slurm_job} object.
}
\description{
Set memory for a Slurm job
}
\examples{
job <- slurm_job("echo Hello") \%>\% set_mem("16G")
job <- set_mem(job, "32G")  # alternative non-pipe usage

}
