% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_opt}
\alias{set_opt}
\title{Set arbitrary Slurm options on a job}
\usage{
set_opt(job, ..., .list = NULL)
}
\arguments{
\item{job}{A \code{slurm_job} object.}

\item{...}{Named Slurm options to set.}

\item{.list}{An optional named list of Slurm options.}
}
\value{
Updated \code{slurm_job} object.
}
\description{
Allows setting one or more Slurm options that are not covered by
dedicated setter functions (e.g. \code{mail_user}, \code{mail_type},
\code{qos}, cluster-specific flags).
}
\details{
Option names should be provided using underscore-separated R-style
names (e.g. \code{mail_user}), which will automatically be translated
to SBATCH flags (e.g. \code{--mail-user}) when rendering.

This function is fully chainable and intended as an advanced or
escape-hatch interface.
}
\examples{
# Set mail notifications
job <- slurm_job("echo Hello") \%>\%
  set_opt(mail_user = "me@lab.edu", mail_type = "END,FAIL")

# Set a cluster-specific option
job <- slurm_job("echo Custom") \%>\%
  set_opt(qos = "high")

# Supply options via a list
opts <- list(
  mail_user = "me@lab.edu",
  mail_type = "BEGIN,END"
)

job <- slurm_job("echo From list") \%>\%
  set_opt(.list = opts)

}
