% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_partition}
\alias{set_partition}
\title{Set Slurm partition}
\usage{
set_partition(job, partition)
}
\arguments{
\item{job}{A \code{slurm_job} object.}

\item{partition}{Partition name.}
}
\value{
Updated \code{slurm_job} object.
}
\description{
Set Slurm partition
}
\examples{
job <- slurm_job("echo Hello") \%>\% set_partition("short")

}
