% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_time}
\alias{set_time}
\title{Set walltime for a Slurm job}
\usage{
set_time(job, time)
}
\arguments{
\item{job}{A \code{slurm_job} object.}

\item{time}{Time limit (e.g. "01:00:00").}
}
\value{
Updated \code{slurm_job} object.
}
\description{
Set walltime for a Slurm job
}
\examples{
job <- slurm_job("echo Hello") \%>\% set_time("02:00:00")

}
